backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4

# Authentication & Security
djangorestframework-simplejwt==5.3.0
python-decouple==3.8

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Kaggle Integration
kaggle==1.5.16

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development
django-extensions==3.2.3
